requirements.txt
django 
djangorestframework 
channels 
channels_redis
